Added component Item
