+++ src/presentation/components/Recipes/components/Item/styles.scss
@@ -1,0 +1,14 @@
+.item {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    &__link {
+        display: block;
+        text-align: center;
+    }
+
+    &__image {
+        display: block;
+    }
+}
